Refactor call_salesforce and exception_handler
With the addition of the bulk API file there was duplicates of
call_salesforce and exception_handler in the file. This moves them to
the util.py file.
