+++ simple_salesforce/tests/test_api.py
@@ -13,28 +13,20 @@ import responses
 try:
     # Python 2.6/2.7
     import httplib as http
-    from mock import Mock, patch
+    from mock import patch
 except ImportError:
     # Python 3
     import http.client as http
-    from unittest.mock import Mock, patch
+    from unittest.mock import patch
 
 import requests
 
 from simple_salesforce import tests
 from simple_salesforce.api import (
-    _exception_handler,
     Salesforce,
     SFType
 )
-from simple_salesforce.exceptions import (
-    SalesforceMoreThanOneRecord,
-    SalesforceMalformedRequest,
-    SalesforceExpiredSession,
-    SalesforceRefusedRequest,
-    SalesforceResourceNotFound,
-    SalesforceGeneralError
-)
+
 
 
 def _create_sf_type(
@@ -568,72 +560,3 @@ class TestSalesforce(unittest.TestCase):
                 instance_url=tests.SERVER_URL, session=session, proxies={})
             self.assertIn('ignoring proxies', mock_log.call_args[0][0])
             self.assertIs(tests.PROXIES, client.session.proxies)
-
-
-class TestExceptionHandler(unittest.TestCase):
-    """Test the exception router"""
-    def setUp(self):
-        """Setup the exception router tests"""
-        self.mockresult = Mock()
-        self.mockresult.url = 'http://www.example.com/'
-        self.mockresult.json.return_value = 'Example Content'
-
-    def test_multiple_records_returned(self):
-        """Test multiple records returned (a 300 code)"""
-        self.mockresult.status_code = 300
-        with self.assertRaises(SalesforceMoreThanOneRecord) as cm:
-            _exception_handler(self.mockresult)
-
-        self.assertEqual(str(cm.exception), (
-            'More than one record for '
-            'http://www.example.com/. Response content: Example Content'))
-
-    def test_malformed_request(self):
-        """Test a malformed request (400 code)"""
-        self.mockresult.status_code = 400
-        with self.assertRaises(SalesforceMalformedRequest) as cm:
-            _exception_handler(self.mockresult)
-
-        self.assertEqual(str(cm.exception), (
-            'Malformed request '
-            'http://www.example.com/. Response content: Example Content'))
-
-    def test_expired_session(self):
-        """Test an expired session (401 code)"""
-        self.mockresult.status_code = 401
-        with self.assertRaises(SalesforceExpiredSession) as cm:
-            _exception_handler(self.mockresult)
-
-        self.assertEqual(str(cm.exception), (
-            'Expired session for '
-            'http://www.example.com/. Response content: Example Content'))
-
-    def test_request_refused(self):
-        """Test a refused request (403 code)"""
-        self.mockresult.status_code = 403
-        with self.assertRaises(SalesforceRefusedRequest) as cm:
-            _exception_handler(self.mockresult)
-
-        self.assertEqual(str(cm.exception), (
-            'Request refused for '
-            'http://www.example.com/. Response content: Example Content'))
-
-    def test_resource_not_found(self):
-        """Test resource not found (404 code)"""
-        self.mockresult.status_code = 404
-        with self.assertRaises(SalesforceResourceNotFound) as cm:
-            _exception_handler(self.mockresult, 'SpecialContacts')
-
-        self.assertEqual(str(cm.exception), (
-            'Resource SpecialContacts Not'
-            ' Found. Response content: Example Content'))
-
-    def test_generic_error_code(self):
-        """Test an error code that is otherwise not caught"""
-        self.mockresult.status_code = 500
-        with self.assertRaises(SalesforceGeneralError) as cm:
-            _exception_handler(self.mockresult)
-
-        self.assertEqual(str(cm.exception), (
-            'Error Code 500. Response content'
-            ': Example Content'))
