+++ simple_salesforce/util.py
@@ -2,6 +2,12 @@
 
 import xml.dom.minidom
 
+from simple_salesforce.exceptions import (
+    SalesforceGeneralError, SalesforceExpiredSession,
+    SalesforceMalformedRequest, SalesforceMoreThanOneRecord,
+    SalesforceRefusedRequest, SalesforceResourceNotFound
+)
+
 
 # pylint: disable=invalid-name
 def getUniqueElementValueFromXmlString(xmlString, elementName):
@@ -33,3 +39,39 @@ def date_to_iso8601(date):
         tzsign=timezone_sign,
         timezone=timezone_str
         ).replace(':', '%3A').replace('+', '%2B')
+
+
+def exception_handler(result, name=""):
+    """Exception router. Determines which error to raise for bad results"""
+    try:
+        response_content = result.json()
+    # pylint: disable=broad-except
+    except Exception:
+        response_content = result.text
+
+    exc_map = {
+        300: SalesforceMoreThanOneRecord,
+        400: SalesforceMalformedRequest,
+        401: SalesforceExpiredSession,
+        403: SalesforceRefusedRequest,
+        404: SalesforceResourceNotFound,
+    }
+    exc_cls = exc_map.get(result.status_code, SalesforceGeneralError)
+
+    raise exc_cls(result.url, result.status_code, name, response_content)
+
+
+def call_salesforce(url, method, session, headers, **kwargs):
+    """Utility method for performing HTTP call to Salesforce.
+
+    Returns a `requests.result` object.
+    """
+
+    additional_headers = kwargs.pop('additional_headers', dict())
+    headers.update(additional_headers or dict())
+    result = session.request(method, url, headers=headers, **kwargs)
+
+    if result.status_code >= 300:
+        exception_handler(result)
+
+    return result
